# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Rawmart Blog API",
        "description": "API documentation for the RAWMART BLOG ASSESSMENT API",
        "contact": {
            "email": "ibrahimsharafadeen95@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/posts/feeds": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Get public feeds",
                "operationId": "a009a0b2587c3522b6e0edcd7da41319",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated list of posts",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PaginatedPosts"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/posts/personal": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Get personal posts",
                "operationId": "cf7c2e88e55ab6c6f077a08222130b1e",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated list of personal posts",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PaginatedPosts"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/posts/{postId}": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Get a single post",
                "operationId": "78b44dcf3b1465052ba67bdf04f5acb5",
                "parameters": [
                    {
                        "name": "postId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Post details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Post"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Post not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Posts"
                ],
                "summary": "Update a post",
                "operationId": "a1ffb6a31b3f9ac65e6cdda21d214484",
                "parameters": [
                    {
                        "name": "postId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Updated title"
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "Updated content"
                                    },
                                    "tags": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "example": "frontend"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Post updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Post"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Post not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Posts"
                ],
                "summary": "Delete a post",
                "operationId": "5adb03a71d3daefafc56251bd135bf13",
                "parameters": [
                    {
                        "name": "postId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Post deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Post deleted"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Post not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/posts": {
            "post": {
                "tags": [
                    "Posts"
                ],
                "summary": "Create a new post",
                "operationId": "fda7c445c718e683207869b051fc183b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "content",
                                    "tags"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "The Future of React 19"
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "Exploring the new compiler and actions API..."
                                    },
                                    "tags": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "example": "react"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Post created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Post"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "UserSummary": {
                "description": "Minimal user representation",
                "properties": {
                    "id": {
                        "type": "string",
                        "example": "new-user"
                    },
                    "name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "image": {
                        "type": "string",
                        "example": "https://images.unsplash.com/photo-1535713875002-d1d0cf377fde"
                    }
                },
                "type": "object"
            },
            "Comment": {
                "description": "Comment resource representation",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 101
                    },
                    "time": {
                        "type": "string",
                        "example": "3 minutes ago"
                    },
                    "user_id": {
                        "type": "string",
                        "example": "new-user"
                    },
                    "text": {
                        "type": "string",
                        "example": "Great insights!"
                    },
                    "user": {
                        "$ref": "#/components/schemas/UserSummary"
                    }
                },
                "type": "object"
            },
            "Post": {
                "description": "Post resource representation",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "The Future of React 19"
                    },
                    "content": {
                        "type": "string",
                        "example": "Exploring the new compiler and actions API in the upcoming release..."
                    },
                    "tags": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "example": "react"
                        }
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-12-24T06:36:00.000Z"
                    },
                    "user": {
                        "$ref": "#/components/schemas/UserSummary"
                    },
                    "comments": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Comment"
                        }
                    }
                },
                "type": "object"
            },
            "PaginationMeta": {
                "description": "Pagination metadata",
                "properties": {
                    "current_page": {
                        "type": "integer",
                        "example": 1
                    },
                    "last_page": {
                        "type": "integer",
                        "example": 5
                    },
                    "per_page": {
                        "type": "integer",
                        "example": 10
                    },
                    "total": {
                        "type": "integer",
                        "example": 50
                    }
                },
                "type": "object"
            },
            "PaginationLinks": {
                "description": "Pagination navigation links",
                "properties": {
                    "first": {
                        "type": "string",
                        "example": "http://api.test/api/posts?page=1"
                    },
                    "last": {
                        "type": "string",
                        "example": "http://api.test/api/posts?page=5"
                    },
                    "prev": {
                        "type": "string",
                        "example": "http://api.test/api/posts?page=1",
                        "nullable": true
                    },
                    "next": {
                        "type": "string",
                        "example": "http://api.test/api/posts?page=3",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "PaginatedPosts": {
                "description": "Paginated response of posts",
                "properties": {
                    "status": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "Success"
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Post"
                        }
                    },
                    "meta": {
                        "$ref": "#/components/schemas/PaginationMeta"
                    },
                    "links": {
                        "$ref": "#/components/schemas/PaginationLinks"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
}